// src/openapi-types.ts
export interface components {
  schemas: {
    Order: {
      id?: number;
      petId?: number;
      quantity?: number;
      shipDate?: string;
      /** Order Status */
      status?: "placed" | "approved" | "delivered";
      complete?: boolean;
    };
    Customer: {
      id?: number;
      username?: string;
      address?: components["schemas"]["Address"][];
    };
    Address: {
      street?: string;
      city?: string;
      state?: string;
      zip?: string;
    };
    Category: {
      id?: number;
      name?: string;
    };
    User: {
      id?: number;
      username?: string;
      firstName?: string;
      lastName?: string;
      email?: string;
      password?: string;
      phone?: string;
      /** User Status */
      userStatus?: number;
    };
    Tag: {
      id?: number;
      name?: string;
    };
    Pet: {
      id?: number;
      name: string;
      category?: components["schemas"]["Category"];
      photoUrls: string[];
      tags?: components["schemas"]["Tag"][];
      /** pet status in the store */
      status?: "available" | "pending" | "sold";
    };
    ApiResponse: {
      code?: number;
      type?: string;
      message?: string;
    };
  };
  requestBodies: {
    /** Pet object that needs to be added to the store */
    Pet: {
      content: {
        "application/json": components["schemas"]["Pet"];
      };
    };
    /** List of user object */
    UserArray: {
      content: {
        "application/json": components["schemas"]["User"][];
      };
    };
  };
}